refactor: move UserIcon
